requirements.txt
google_auth_oauthlib==0.4.6
google_api_python_client==2.37.0
learnvcs==0.3.6
protobuf==3.19.4
